+++ omnibenchmark/cli/run.py
@@ -1,9 +1,6 @@
 """cli commands related to benchmark/module execution and start"""
 
-import os
 import sys
-import tempfile
-from itertools import chain
 from pathlib import Path
 
 import click
@@ -19,32 +16,12 @@ from omnibenchmark.remote.storage import remote_storage_snakemake_args
 from omnibenchmark.workflow.snakemake import SnakemakeEngine
 from omnibenchmark.workflow.workflow import WorkflowEngine
 
-from .debug import add_debug_option
 
-
-@click.group(name="run")
-@click.pass_context
-def run(ctx):
-    """Run benchmarks or benchmark modules."""
-    ctx.ensure_object(dict)
-
-
-@add_debug_option
-@run.command(
-    name="benchmark",
-    context_settings=dict(
-        allow_extra_args=True,
-    ),
+@click.command(
+    name="run",
+    context_settings=dict(allow_extra_args=True),
 )
-@click.option(
-    "-b",
-    "--benchmark",
-    "benchmark_path",
-    help="Path to benchmark yaml file or benchmark id.",
-    required=True,
-    envvar="OB_BENCHMARK",
-    type=click.Path(exists=True),
-)
+@click.argument("benchmark_path", type=click.Path(exists=True))
 @click.option(
     "-c",
     "--cores",
@@ -105,7 +82,7 @@ def run(ctx):
     type=str,
 )
 @click.pass_context
-def run_benchmark(
+def run(
     ctx,
     benchmark_path,
     cores,
@@ -193,226 +170,6 @@ def run_benchmark(
     )
 
     log_result_and_quit(logger, success, type="Benchmark")
-
-
-@run.command(
-    name="module",
-    context_settings=dict(
-        allow_extra_args=True,
-    ),
-)
-@click.option(
-    "-b",
-    "--benchmark",
-    "benchmark_path",
-    help="Path to benchmark yaml file or benchmark id.",
-    required=True,
-    envvar="OB_BENCHMARK",
-    type=click.Path(exists=True),
-)
-@click.option("-m", "--module", help="Module id to execute", type=str, required=True)
-@click.option(
-    "-i",
-    "--input_dir",
-    help="Path to the folder with the appropriate input files.",
-    type=click.Path(exists=True, writable=True),
-    default=None,
-)
-@click.option("-d", "--dry", help="Dry run.", is_flag=True, default=False)
-@click.option(
-    "-u",
-    "--update",
-    help="Force re-run execution for all modules and stages.",
-    is_flag=True,
-    default=False,
-)
-@click.option(
-    "-k",
-    "--continue-on-error",
-    help="Go on with independent jobs if a job fails (--keep-going in snakemake).",
-    is_flag=True,
-    default=False,
-)
-@click.option(
-    "--keep-module-logs/--no-keep-module-logs",
-    default=False,
-    help="Keep module-specific log files after execution.",
-)
-@click.option(
-    "--task-timeout",
-    type=str,
-    default=None,
-    help="A `human friendly` timeout for each separate task execution (local only). Do note that total runtime is not additive. Example: 4h, 42m, 12s",
-)
-@click.option(
-    "--executor",
-    help="Specify a custom executor to use for workflow execution. Example: slurm",
-    type=str,
-    default=None,
-)
-@click.pass_context
-def run_module(
-    ctx,
-    benchmark_path,
-    module,
-    input_dir,
-    dry,
-    update,
-    keep_module_logs,
-    continue_on_error,
-    task_timeout,
-    executor,
-):
-    """
-    Run a specific module that is part of the benchmark.
-    """
-    behaviours = {"input": input_dir, "example": None, "all": None}
-    extra_args = parse_extra_args(ctx.args)
-
-    if task_timeout is None:
-        timeout_s = None
-    else:
-        try:
-            timeout_s = int(humanfriendly.parse_timespan(task_timeout))
-        except humanfriendly.InvalidTimespan:
-            logger.error(f"Invalid timeout value: {task_timeout}")
-            sys.exit(1)
-
-    non_none_behaviours = {
-        key: value for key, value in behaviours.items() if value is not None
-    }
-    if len(non_none_behaviours) >= 2:
-        log_error_and_quit(
-            logger,
-            "Error: Only one of '--input_dir', '--example', or '--all' should be set. Please choose only one option.",
-        )
-        return
-
-    # Construct a message specifying which option is set
-    behaviour = list(non_none_behaviours)[0] if non_none_behaviours else None
-
-    if behaviour == "example" or behaviour == "all":
-        howmany = "all" if behaviour == "all" else "a"
-        logger.info(f"Running module on {howmany} predefined remote example dataset.")
-
-        # TODO Check how snakemake storage decorators work, do we have caching locally or just remote?
-        # TODO Implement remote execution using remote url from benchmark definition
-        log_error_and_quit(
-            logger,
-            "Error: Remote execution is not supported yet. Workflows can only be run in local mode.",
-        )
-        return
-
-    logger.info("Running module on a local dataset.")
-
-    try:
-        b = BenchmarkExecution(Path(benchmark_path), Path(tempfile.mkdtemp()))
-    except BenchmarkParseError as e:
-        formatted_error = pretty_print_parse_error(e)
-        log_error_and_quit(logger, f"Failed to load benchmark: {formatted_error}")
-        return
-    except Exception as e:
-        log_error_and_quit(logger, f"Failed to load benchmark: {e}")
-        return
-    assert b is not None
-    benchmark_nodes = b.get_nodes_by_module_id(module_id=module)
-
-    is_entrypoint_module = all([node.is_entrypoint() for node in benchmark_nodes])
-
-    if len(benchmark_nodes) <= 0:
-        log_error_and_quit(
-            logger,
-            f"Error: Could not find module with id `{module}` in benchmark definition",
-        )
-        return
-
-    assert len(benchmark_nodes) > 0
-    logger.info(f"Found {len(benchmark_nodes)} workflow nodes for module {module}.")
-
-    if not is_entrypoint_module and len(non_none_behaviours) == 0:
-        log_error_and_quit(
-            logger,
-            "Error: At least one option must be specified. Use '--input_dir', '--example', or '--all'.",
-        )
-        return
-
-    if input_dir is None:
-        input_dir = os.getcwd()
-
-    logger.info("Running module benchmark...")
-
-    # Check if input path exists and is a directory
-    if not (os.path.exists(input_dir) and os.path.isdir(input_dir)):
-        log_error_and_quit(
-            logger,
-            f"Error: Input directory does not exist or is not a valid directory: `{input_dir}`",
-        )
-        return
-    assert os.path.exists(input_dir) and os.path.isdir(input_dir)
-
-    benchmark_datasets = b.get_benchmark_datasets()
-
-    # Check available files in input to figure out what dataset are we processing
-    if module in benchmark_datasets:
-        # we're given the initial dataset module to process, then we know
-        dataset = module
-    else:
-        # we try to figure the dataset based on the files present in the input directory
-        files = os.listdir(input_dir)
-        base_names = [file.split(".")[0] for file in files]
-        dataset = next((d for d in benchmark_datasets if d in base_names), None)
-
-    if dataset is None:
-        log_error_and_quit(
-            logger,
-            f"Error: Could not infer the appropriate dataset to run the node workflow on based on the files available in `{input_dir}`. None of the available datasets {benchmark_datasets} match the base names of the files.",
-        )
-        return
-
-    assert dataset is not None
-
-    # Check if input directory contains all necessary input files
-    required_inputs = list(map(lambda node: node.get_inputs(), benchmark_nodes))
-    required_inputs = list(chain.from_iterable(required_inputs))
-    required_input_files = list(
-        set([os.path.basename(path) for path in required_inputs])
-    )
-    required_input_files = [
-        file.format(dataset=dataset) for file in required_input_files
-    ]
-
-    input_files = os.listdir(input_dir)
-    missing_files = [file for file in required_input_files if file not in input_files]
-
-    if len(missing_files) > 0:
-        log_error_and_quit(
-            logger,
-            f"Error: The following required input files are missing from the input directory: {missing_files}.",
-        )
-        return
-    assert len(missing_files) == 0
-
-    workflow: WorkflowEngine = SnakemakeEngine()
-
-    for benchmark_node in benchmark_nodes:
-        # When running a single module, it doesn't have sense to make parallelism level (cores) configurable
-        success = workflow.run_node_workflow(
-            node=benchmark_node,
-            input_dir=Path(input_dir),
-            dataset=dataset,
-            cores=1,
-            update=update,
-            dryrun=dry,
-            continue_on_error=continue_on_error,
-            keep_module_logs=keep_module_logs,
-            backend=b.get_benchmark_software_backend(),
-            executor=executor,
-            local_timeout=timeout_s,
-            benchmark_file_path=b.get_definition_file(),
-            **extra_args,
-        )
-
-        log_result_and_quit(logger, success, type="Module")
 
 
 def log_error_and_quit(logger, error):